Move worker pool out of src/client
