+++ src/server/pkg/pool/pool.go
@@ -1,4 +1,4 @@
-package grpcutil
+package pool
 
 import (
 	"context"
